Move generic code from SAT specific file to shared telescope file and fix some typing issues
