+++ src/schedlib/policies/satp1.py
@@ -3,7 +3,8 @@ from dataclasses import dataclass
 import datetime as dt
 
 from .. import source as src, utils as u
-from .sat import SATPolicy, State, CalTarget
+from .sat import SATPolicy, State
+from .tel import CalTarget
 from ..commands import SchedMode
 
 logger = u.init_logger(__name__)
